requirements.txt
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
Django==1.11.20
django-annoying==0.10.4
django-appconf==1.0.2
django-bower==5.2.0
django-contrib-comments==1.9.0
django-extensions==2.2.6
django-grappelli==2.12.2
django-happenings==0.3.3
django-hitcount==1.3.1
django-imagekit==4.0.2
django-jquery==3.1.0
django-slugify-processor==0.8.4
docopt==0.6.2
entrypoints==0.3
filebrowser-safe==0.5.0
flake8==3.7.5
future==0.17.1
grappelli-safe==0.5.1
html5lib==1.0.1
icalendar==4.0.3
idna==2.8
isort==4.3.4
mccabe==0.6.1
Mezzanine==4.3.1
mock==2.0.0
oauthlib==3.0.1
olefile==0.46
pbr==5.1.2
pilkit==2.0
Pillow==5.4.1
pycodestyle==2.5.0
pyflakes==2.1.0
python-dateutil==2.8.0
python-twitter==3.5
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.8
sqlparse==0.3.0
tzlocal==1.5.1
urllib3==1.24.1
webencodings==0.5.1
